moved antivirusInfo
