--- packages/web-app/src/modules/onboarding-views/AntivirusInfo.ts
@@ -1,185 +1,0 @@
-import { AntiVirusSoftware } from '../zendesk/models'
-import Adaware from './assets/Adaware.png'
-import AdvSystemCare from './assets/AdvSystemCare.png'
-import Avast from './assets/Avast.png'
-import AVG from './assets/AVG.png'
-import Avira from './assets/Avira.png'
-import BitDefender from './assets/BitDefender.png'
-import BullGuard from './assets/BullGuard.png'
-import Comodo from './assets/Comodo.png'
-import EScan from './assets/EScan.png'
-import EsetNod32 from './assets/EsetNod32.png'
-import FProt from './assets/FProt.png'
-import FSecure from './assets/FSecure.png'
-import GData from './assets/GData.png'
-import K7 from './assets/K7.png'
-import Kaspersky from './assets/Kaspersky.png'
-import Malwarebytes from './assets/Malwarebytes.png'
-import McAfeeSecurity from './assets/McAfeeSecurity.png'
-import Norton from './assets/Norton.png'
-import Panda from './assets/Panda.png'
-import PCMatic from './assets/PCMatic.png'
-import Qihoo360 from './assets/Qihoo360.png'
-import SecureAgeAPEX from './assets/SecureAgeAPEX.png'
-import Sophos from './assets/Sophos.png'
-import TotalAV from './assets/TotalAV.png'
-import TrendMicro from './assets/TrendMicro.png'
-import Twister from './assets/Twister.png'
-import VIPRE from './assets/VIPRE.png'
-import Webroot from './assets/Webroot.png'
-import WindowsDefender from './assets/WindowsDefender.png'
-import Zillya from './assets/Zillya.png'
-
-export const antivirusInfo = [
-  {
-    label: 'Windows Defender',
-    name: AntiVirusSoftware.WindowsDefender,
-    src: WindowsDefender,
-  },
-  {
-    label: 'Avast',
-    name: AntiVirusSoftware.Avast,
-    src: Avast,
-  },
-  {
-    label: 'McAfee',
-    name: AntiVirusSoftware.McAfeeSecurity,
-    src: McAfeeSecurity,
-  },
-  {
-    label: 'Norton',
-    name: AntiVirusSoftware.Norton,
-    src: Norton,
-  },
-
-  {
-    label: 'Malwarebytes',
-    name: AntiVirusSoftware.Malwarebytes,
-    src: Malwarebytes,
-  },
-  {
-    label: 'Kaspersky',
-    name: AntiVirusSoftware.Kaspersky,
-    src: Kaspersky,
-  },
-  {
-    label: 'Avira',
-    name: AntiVirusSoftware.Avira,
-    src: Avira,
-  },
-  {
-    label: 'BitDefender',
-    name: AntiVirusSoftware.BitDefender,
-    src: BitDefender,
-  },
-  {
-    label: 'AVG',
-    name: AntiVirusSoftware.AVG,
-    src: AVG,
-  },
-  {
-    label: 'ESET-NOD32',
-    name: AntiVirusSoftware.ESETNOD32,
-    src: EsetNod32,
-  },
-  {
-    label: 'Webroot',
-    name: AntiVirusSoftware.Webroot,
-    src: Webroot,
-  },
-  {
-    label: 'TotalAV',
-    name: AntiVirusSoftware.TotalAV,
-    src: TotalAV,
-  },
-  {
-    label: 'VIPRE',
-    name: AntiVirusSoftware.VIPRE,
-    src: VIPRE,
-  },
-  {
-    label: 'TrendMicro',
-    name: AntiVirusSoftware.TrendMicro,
-    src: TrendMicro,
-  },
-  {
-    label: 'Bullguard',
-    name: AntiVirusSoftware.BullGuard,
-    src: BullGuard,
-  },
-  {
-    label: 'F-Secure',
-    name: AntiVirusSoftware.FSecure,
-    src: FSecure,
-  },
-  {
-    label: 'Panda',
-    name: AntiVirusSoftware.Panda,
-    src: Panda,
-  },
-  {
-    label: 'K7',
-    name: AntiVirusSoftware.K7,
-    src: K7,
-  },
-  {
-    label: 'Zillya',
-    name: AntiVirusSoftware.Zillya,
-    src: Zillya,
-  },
-  {
-    label: 'Sophos',
-    name: AntiVirusSoftware.Sophos,
-    src: Sophos,
-  },
-  {
-    label: 'Twister',
-    name: AntiVirusSoftware.Twister,
-    src: Twister,
-  },
-  {
-    label: 'GData',
-    name: AntiVirusSoftware.GData,
-    src: GData,
-  },
-  {
-    label: 'PCMatic',
-    name: AntiVirusSoftware.PCMatic,
-    src: PCMatic,
-  },
-  {
-    label: 'Qihoo-360',
-    name: AntiVirusSoftware.Qihoo360,
-    src: Qihoo360,
-  },
-  {
-    label: 'Adv. SystemCare',
-    name: AntiVirusSoftware.AdvancedSystemCareUltimate,
-    src: AdvSystemCare,
-  },
-  {
-    label: 'Adaware',
-    name: AntiVirusSoftware.Adaware,
-    src: Adaware,
-  },
-  {
-    label: 'e-Scan',
-    name: AntiVirusSoftware.EScan,
-    src: EScan,
-  },
-  {
-    label: 'F-Prot',
-    name: AntiVirusSoftware.FProt,
-    src: FProt,
-  },
-  {
-    label: 'SecureAge APEX',
-    name: AntiVirusSoftware.SecureAgeAPEX,
-    src: SecureAgeAPEX,
-  },
-  {
-    label: 'Comodo',
-    name: AntiVirusSoftware.Comodo,
-    src: Comodo,
-  },
-]
